requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.0
nltk==3.8.1
textstat==0.7.3
python-dateutil==2.9.0.post0
pyyaml==6.0.2
flask==3.0.3
python-docx==1.1.2
pdfminer.six==20240706
joblib==1.4.2
